spos_starter/spos_starter/requirements.txt
exa-py
firecrawl-py
playwright
python-dotenv
